[feat] Support reduction on the induction variable. [fix] Fix a bug in removing ExitValue user for reduction stream.
1. To support reduction like:
if (a[i] == x) sum += i;
We need to relax the constraint on the input of the ReduceDG. Specifically,
we allow it takes a LoadStream and its StepRootStream, which is "i", as inputs.

2. Originally, when we remove an user instruction of the ExitValue of a reduction stream,
we erase it at the same time when we traverse all the users. However this "modify-while-iterating"
is not supported in LLVM and causes a bug.

To fix this, we first collect all out-of-loop ExitValue users, and then remove them.